requirements.txt
packaging
wheel
torch
huggingface-hub
transformers
librosa
flash-attn
laion_clap
x-transformers
wandb

argbind>=0.3.7
descript-audiotools>=0.7.2
einops
numpy
torchaudio
tqdm
tensorboard
numba>=0.5.7
jupyterlab
lightning

encodec
torchmetrics
git+https://github.com/WaveGenAI/SemantiCodec-inference.git
git+https://github.com/WaveGenAI/descript-audio-codec.git